refactor: replace all `styles/emotion` usage
