+++ src/components/MdxComponents/Paragraph.ts
@@ -1,4 +1,4 @@
-import { styled } from 'styles/emotion';
+import { styled } from 'styles/styled';
 
 export const Paragraph = styled.p`
   cursor: text;
