+++ src/components/MdxComponents/YouTubeVideo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { styled } from 'styles/emotion';
+import { styled } from 'styles/styled';
 
 const ResponsiveIframe = styled.div`
   margin-bottom: 2rem;
