+++ src/components/SideMenu/styles.ts
@@ -1,4 +1,4 @@
-import { styled, css } from 'styles/emotion';
+import { styled, css } from 'styles/styled';
 import { Link } from 'gatsby';
 import { motion } from 'framer-motion';
