+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormattedMessage, defineMessages } from 'react-intl';
 
-import { styled } from 'styles/emotion';
+import { styled } from 'styles/styled';
 import Layout from 'components/Layout';
 import SEO from 'components/SEO';
 import { useIntl } from 'context/react-intl';
